Opencv requirement issue fixed

+++ api.py
@@ -1,9 +1,32 @@
+from PIL import Image
+import numpy as np
+from fastapi import FastAPI, UploadFile, File
+from io import BytesIO
 import tensorflow as tf
-import keras
-import numpy as np 
+import keras 
+import os 
 import cv2
 
-# Function the cut the image to only the eye fundus
+app = FastAPI()
+
+class_names = ['cataract', 'normal']
+model = keras.models.load_model("catnet.h5")
+
+def read_imagefile(file) -> Image.Image:
+    image = Image.open(BytesIO(file))
+    return image
+
+def predict(img_path):  
+    img = tf.keras.utils.load_img(img_path, target_size=(256, 256))
+    img_array = tf.keras.utils.img_to_array(img)
+    img_array = tf.expand_dims(img_array, 0)
+    predictions = model.predict(img_array)
+    score = predictions[0]
+    # score = 100 * np.max(score)
+    res = "This image most likely belongs to {} with a {:.2f} percent confidence.".format(class_names[np.argmax(score)], 100 * np.max(score))
+    res = {"{}".format(class_names[np.argmax(score)], 100 * np.max(score)):"{:.4f}".format(100 * np.max(score))}
+    return res
+
 def thresh_crop_image(img) :
     path = os.path.abspath(img)
     name = os.path.basename(img)
@@ -19,51 +42,26 @@ def thresh_crop_image(img) :
     return path
 
 
-class_names = ['cataract', 'normal']
-model = keras.models.load_model("catnet.h5")
-
-def predict(img_path):  
-    img = tf.keras.utils.load_img(img_path, target_size=(256, 256))
-    img_array = tf.keras.utils.img_to_array(img)
-    img_array = tf.expand_dims(img_array, 0)
-    predictions = model.predict(img_array)
-    score = predictions[0]
-    # score = 100 * np.max(score)
-    res = "This image most likely belongs to {} with a {:.2f} percent confidence.".format(class_names[np.argmax(score)], 100 * np.max(score))
-    res = {"{}".format(class_names[np.argmax(score)], 100 * np.max(score)):"{:.4f}".format(100 * np.max(score))}
-    return res
-
-
-from fastapi import FastAPI, File, UploadFile
-import uuid
-import os 
-
-app = FastAPI()
-
 IMAGEDIR = "images/"
 if not os.path.exists(IMAGEDIR) :
     os.mkdir(IMAGEDIR)
 
-
 @app.get('/')
-def root():
-    return {"Response":"Working Great!"}
+async def root():
+    return {"Response":"Wokring Great!"}
 
+@app.post("/predict/")
+async def predict_api(file: UploadFile = File(...)):
+    extension = file.filename.split(".")[-1] in ("jpg", "jpeg", "png")
+    if not extension:
+        return "Image must be jpg, jpeg, or png format!"
 
-@app.post('/diagnose')
-async def create_upload_file(file: UploadFile = File(...)):
- 
-    file.filename = f"{uuid.uuid4()}.jpg"
     contents = await file.read()
- 
-    #save the file
     with open(f"{IMAGEDIR}{file.filename}", "wb") as f:
         f.write(contents)
 
     file_path = f"{IMAGEDIR}{file.filename}"
-
-    thresed_and_cropped = thresh_crop_image(file_path)
-    diagnosis = predict(thresed_and_cropped)
+    threshed = thresh_crop_image(file_path)
+    prediction = predict(threshed)
     os.remove(file_path)
-
-    return diagnosis
+    return prediction
